refactor: clean up upgrade path
