+++ cyanprint/src/run.rs
@@ -5,6 +5,9 @@ use std::rc::Rc;
 
 use cyancoordinator::client::CyanCoordinatorClient;
 use cyancoordinator::fs::{DiskFileLoader, DiskFileWriter, GitLikeMerger, TarGzUnpacker};
+use cyancoordinator::operations::composition::{
+    CompositionOperator, DefaultDependencyResolver, DefaultVfsLayerer,
+};
 use cyancoordinator::operations::{TemplateOperations, TemplateOperator};
 use cyanregistry::http::client::CyanRegistryClient;
 use cyanregistry::http::models::template_res::TemplateVersionRes;
@@ -14,7 +17,7 @@ use cyancoordinator::session::SessionIdGenerator;
 use cyancoordinator::template::DefaultTemplateExecutor;
 use cyancoordinator::template::{DefaultTemplateHistory, TemplateHistory, TemplateUpdateType};
 
-/// Run the cyan template generation process
+/// Run the cyan template generation process with automatic composition detection
 /// Returns all session IDs that were created and need to be cleaned up
 pub fn cyan_run(
     session_id_generator: Box<dyn SessionIdGenerator>,
@@ -25,7 +28,7 @@ pub fn cyan_run(
     registry_client: Rc<CyanRegistryClient>,
     debug: bool,
 ) -> Result<Vec<String>, Box<dyn Error + Send>> {
-    // handle the target directory
+    // Handle the target directory
     let path = path.unwrap_or(".".to_string());
     let path_buf = PathBuf::from(&path);
     let target_dir = path_buf.as_path();
@@ -33,64 +36,106 @@ pub fn cyan_run(
     fs::create_dir_all(target_dir).map_err(|e| Box::new(e) as Box<dyn Error + Send>)?;
 
     // Create all components for dependency injection at the highest level
-    let unpacker: Box<dyn cyancoordinator::fs::FileUnpacker> = Box::new(TarGzUnpacker);
-    let loader: Box<dyn cyancoordinator::fs::FileLoader> = Box::new(DiskFileLoader);
-    let merger: Box<dyn cyancoordinator::fs::FileMerger> = Box::new(GitLikeMerger::new(debug, 50));
-    let writer: Box<dyn cyancoordinator::fs::FileWriter> = Box::new(DiskFileWriter);
+    let unpacker = Box::new(TarGzUnpacker);
+    let loader = Box::new(DiskFileLoader);
+    let merger = Box::new(GitLikeMerger::new(debug, 50));
+    let writer = Box::new(DiskFileWriter);
 
     // Setup services with explicit dependencies
     let template_history = Box::new(DefaultTemplateHistory::new());
     let template_executor = Box::new(DefaultTemplateExecutor::new(coord_client.endpoint.clone()));
     let vfs = Box::new(DefaultVfs::new(unpacker, loader, merger, writer));
 
-    // Create the TemplateOperator with all dependencies including registry client
+    // Create the TemplateOperator
     let template_operator = TemplateOperator::new(
         session_id_generator,
         template_executor,
         template_history,
         vfs,
-        registry_client,
+        registry_client.clone(),
     );
 
     // Check template history to determine update scenario
     let update_type =
         DefaultTemplateHistory::new().check_template_history(target_dir, &template, &username)?;
 
-    // Handle different update scenarios and collect all session IDs for cleanup
-    match update_type {
-        TemplateUpdateType::NewTemplate => {
-            // Scenario 1: No previous template matching the current template
-            template_operator.create_new(&template, target_dir, &username)
-        }
-        TemplateUpdateType::UpgradeTemplate {
-            previous_version,
-            previous_answers,
-            previous_states,
-        } => {
-            // Scenario 2: Previous template matching the current template exists, but a different version
-            template_operator.upgrade(
+    // Auto-detect if template has dependencies and use appropriate execution path
+    if template.templates.is_empty() {
+        println!("📦 Template has no dependencies - using single template execution");
+
+        // Handle different update scenarios for single templates
+        match update_type {
+            TemplateUpdateType::NewTemplate => {
+                template_operator.create_new(&template, target_dir, &username)
+            }
+            TemplateUpdateType::UpgradeTemplate {
+                previous_version,
+                previous_answers,
+                previous_states,
+            } => template_operator.upgrade(
                 &template,
                 target_dir,
                 &username,
                 previous_version,
                 previous_answers,
                 previous_states,
-            )
-        }
-        TemplateUpdateType::RerunTemplate {
-            previous_version,
-            previous_answers,
-            previous_states,
-        } => {
-            // Scenario 3: Previous template matching the current template exists, with the same version
-            template_operator.rerun(
+            ),
+            TemplateUpdateType::RerunTemplate {
+                previous_version,
+                previous_answers,
+                previous_states,
+            } => template_operator.rerun(
                 &template,
                 target_dir,
                 &username,
                 previous_version,
                 previous_answers,
                 previous_states,
-            )
+            ),
+        }
+    } else {
+        println!(
+            "🔗 Template has {} dependencies - using composition execution",
+            template.templates.len()
+        );
+
+        // Create composition-specific components
+        let dependency_resolver = Box::new(DefaultDependencyResolver::new(registry_client.clone()));
+        let vfs_layerer = Box::new(DefaultVfsLayerer);
+
+        // Create the CompositionOperator
+        let composition_operator =
+            CompositionOperator::new(template_operator, dependency_resolver, vfs_layerer);
+
+        // Handle different update scenarios for compositions
+        match update_type {
+            TemplateUpdateType::NewTemplate => {
+                composition_operator.create_new_composition(&template, target_dir, &username)
+            }
+            TemplateUpdateType::UpgradeTemplate {
+                previous_version,
+                previous_answers,
+                previous_states,
+            } => composition_operator.upgrade_composition(
+                &template,
+                target_dir,
+                &username,
+                previous_version,
+                previous_answers,
+                previous_states,
+            ),
+            TemplateUpdateType::RerunTemplate {
+                previous_version,
+                previous_answers,
+                previous_states,
+            } => composition_operator.rerun_composition(
+                &template,
+                target_dir,
+                &username,
+                previous_version,
+                previous_answers,
+                previous_states,
+            ),
         }
     }
 }
